requirements.txt
aniso8601==7.0.0
awsebcli==3.15.0
Babel==2.7.0
blessed==1.15.0
blinker==1.4
cached-property==1.5.1
cachetools==3.1.1
cement==2.8.2
chardet==3.0.4
Click==7.0
decorator==4.3.0
dialogflow==0.6.0
docker==3.7.2
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
Flask==1.0.3
Flask-BabelEx==0.9.3
Flask-Cors==3.0.8
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.7
Flask-Security==3.0.0
Flask-SocketIO==4.1.0
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
future==0.16.0
google-api-core==1.12.0
google-auth==1.6.3
googleapis-common-protos==1.6.0
grpcio==1.21.1
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
lxml==4.2.5
MarkupSafe==1.1.1
passlib==1.7.1
pathspec==0.5.9
plivo==4.1.3
protobuf==3.8.0
psycopg2-binary==2.8.3
pyasn1==0.4.5
pyasn1-modules==0.2.5
python-engineio==3.8.1
python-socketio==4.1.0
pytz==2019.1
requests==2.20.1
rsa==4.0
semantic-version==2.5.0
six==1.11.0
speaklater==1.3
SQLAlchemy==1.3.5
SQLAlchemy-JSONField==0.8.0
termcolor==1.1.0
texttable==0.9.1
typing==3.7.4
ujson==1.35
urllib3==1.24.1
websocket-client==0.56.0
Werkzeug==0.15.4
WTForms==2.2.1
